deploy/requirements.txt
schedule
configparser
requests
urllib3
pandas
Office365-REST-Python-Client